% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shakalalaboomboom.R
\name{shakalalaboomboom}
\alias{shakalalaboomboom}
\title{create Histogram boxplot and numeric summary}
\usage{
shakalalaboomboom(x)
}
\arguments{
\item{x}{numeric variables}
}
\description{
create Histogram boxplot and numeric summary
}
